Improves specification for replace_first()
This improves the specifications given for the two replace_first() methods.
This required adding various helper properties to keep things relatively tidy.

In general, it makes me wonder a lot about the value of using witnesses.  In
particular, the specifications for replace_first() would be easier to write if
they returned the actual index of a match (or null if no match).