Add initial Option implementation
This adds an initial implementation of option.  This is not strictly necessary
in Whiley, as we can use union types directly for this.  However, it is
particularly useful for use with templates.  That is, when we cannot be sure
whether the template parameter T can be null or not.

Have additionally included a number of potentially helpful functions for
manipulating option types.  It's not clear to me at this stage what other
functions might be useful.