Further improvements to std:array
This further improves the set of methods which have been verified in std::array.
At this stage, I'm stuck on a loop invariant inside copy<T> which feels like it
should verify.  Its possible the translation of for loops is the problem, and
one way forward would be to try a while loop.